fix: break circular dependency in embedded-agent builtin tools (#1042)
Extracts BuiltinTool/BuiltinToolContext/BuiltinToolResult into a new
types.ts so read.ts/glob.ts/grep.ts no longer import back from index.ts
(which imports them). index.ts re-exports the types so existing consumers
(composite-executor.ts, main.ts, tests) are unaffected.
