+++ packages/embedded-agent/src/tools/grep.ts
@@ -7,7 +7,7 @@
 import { Glob } from 'bun';
 import * as fsPromises from 'node:fs/promises';
 import * as path from 'node:path';
-import type { BuiltinTool, BuiltinToolContext, BuiltinToolResult } from './index.js';
+import type { BuiltinTool, BuiltinToolContext, BuiltinToolResult } from './types.js';
 import { resolveConfinedPath } from './path-confinement.js';
 
 /** Pragmatic v1 limit to bound worst-case latency/memory; not a spec requirement. */
